split window plan from aggregate plan
